move helper functions to see if it resolve issue
